auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose
redis
pika
pydantic[email]
